refactor: restructure the API
