+++ api/dependencies.py
@@ -1,0 +1,7 @@
+
+from api.models import WeaviateClient
+
+
+async def get_client():
+    with WeaviateClient() as client:
+        yield client
